Refactors client Logout
