+++ apps/web/src/lib/utils/server/user.ts
@@ -1,0 +1,5 @@
+import { redirect } from "next/navigation";
+export async function clientLogOut() {
+	"use server";
+	redirect("/");
+}
